cloud-router/requirements.txt
Flask
azure-cosmos
folium